pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram",
    "sqlalchemy",
    "pillow",
    "pydantic",
    "pydantic-settings",
    "psycopg[binary]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["handlers*", "db*"]
